requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
APScheduler==3.6.3
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
multidict==4.7.6
python-dotenv==1.0.0
python-telegram-bot==20.4
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tornado==6.3.3
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.4
yarl==1.9.2

aiogram~=3.12.0
Telethon~=1.36.0